+++ system/src/Grav/Common/GPM/Remote/GravCore.php
@@ -21,7 +21,6 @@ class GravCore extends AbstractPackageCollection
 {
     /** @var string */
     protected $repository = 'https://getgrav.org/downloads/grav.json';
-
     /** @var array */
     private $data;
     /** @var string */